aoj/ALDS1/ALDS1_8_B_test.py
from ALDS1_8_B import resolve
import sys
from io import StringIO
import unittest

class TestClass(unittest.TestCase):
    maxDiff = None
    def assertIO(self, input, output):
        stdout, stdin = sys.stdout, sys.stdin
        sys.stdout, sys.stdin = StringIO(), StringIO(input)
        resolve()
        sys.stdout.seek(0)
        out = sys.stdout.read()[:-1]
        sys.stdout, sys.stdin = stdout, stdin
        self.assertEqual(out, output)
    def test_入力例_1(self):
        input = """10
insert 30
insert 88
insert 12
insert 1
insert 20
find 12
insert 17
insert 25
find 16
print"""
        output = """yes
no
 1 12 17 20 25 30 88
 30 12 1 20 17 25 88"""
        self.assertIO(input, output)

if __name__ == "__main__":
    unittest.main()